refactor(registration): refactored registration routes
