+++ client/src/common/styles/NavBar.style.ts
@@ -1,0 +1,46 @@
+import styled from 'styled-components';
+
+export const NavBar = styled.nav`
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+  height: 8vh;
+  width: 100%;
+  a {
+    color: #fff5ee;
+    text-decoration: none;
+    margin-right: 2em;
+  }
+  a:hover {
+    text-decoration: underline;
+  }
+  button {
+    cursor: pointer;
+  }
+  .app-name {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    margin-left: 1em;
+    color: #fff5ee;
+    h1 {
+      font-size: 2rem;
+      margin-right: 0.5em;
+    }
+    img {
+      height: 95%;
+      width: 3em;
+    }
+  }
+  .navigation {
+    button {
+      background: white;
+      border: none;
+      border-radius: 5px;
+      padding: 7px;
+      margin-right: 1em;
+    }
+  }
+`;
+export default {};
