--- client/src/pages/Main.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { Outlet } from 'react-router-dom';
-import GlobalStyles from '../styles/GlobalStyles';
-
-function Main() {
-  return (
-    <>
-      <GlobalStyles />
-      <Outlet />
-    </>
-  );
-}
-
-export default Main;
